Refactor storage components to separate class
