+++ src/main/java/WenJigglyBotException.java
@@ -1,3 +1,12 @@
 public class WenJigglyBotException extends Exception {
+    String s;
 
+    public WenJigglyBotException(String s) {
+        this.s = s;
+    }
+
+    @Override
+    public String toString() {
+        return s;
+    }
 }
